Move all spells to its own files.
